refactor(스프링 카페 2-3 단계 피드백(#PR584)): 기능에 따른 Controller 분리
1. 댓글 출력 : ArticlePrintController -> ReplyPrintController 이동
2. 댓글 삭제 : ArticleDeleteController -> ReplyDeleteController 이동

+++ src/main/java/com/kakao/cafe/application/reply/service/DeleteReplyService.java
@@ -22,7 +22,11 @@ public class DeleteReplyService implements DeleteReplyUseCase {
     }
 
     @Override
-    public void deleteAllRepliesInArticle(int articleId) {
+    public void deleteAllRepliesInArticle(int articleId, String userIdDeletingReplies, UserInfo sessionedUser)
+        throws UnauthenticatedUserException {
+        if (!sessionedUser.getUserId().equals(userIdDeletingReplies)) {
+            throw new UnauthenticatedUserException("인증 오류");
+        }
         deleteReplyPort.deleteAllRepliesInArticle(articleId);
     }
 }
